examples/payapi/ps_common/requirements.txt
Django==3.0.2
djangorestframework==3.11.0
python-dotenv==0.10.5
requests==2.22.0
